+++ service/controller/resource/nodepool/create_terminate_old_workers.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/Azure/azure-sdk-for-go/services/compute/mgmt/2019-07-01/compute"
 	"github.com/Azure/go-autorest/autorest/to"
+	"github.com/giantswarm/errors/tenant"
 	"github.com/giantswarm/microerror"
 	"sigs.k8s.io/cluster-api/util"
 
@@ -40,11 +41,13 @@ func (r *Resource) terminateOldWorkersTransition(ctx context.Context, obj interf
 	}
 
 	tenantClusterK8sClient, err := r.tenantClientFactory.GetClient(ctx, cluster)
-	if err != nil {
-		r.Logger.LogCtx(ctx, "level", "debug", "message", "tenant API not available yet", "stack", microerror.JSON(err))
+	if tenant.IsAPINotAvailable(err) {
+		r.Logger.LogCtx(ctx, "level", "debug", "message", "tenant API not available yet")
 		r.Logger.LogCtx(ctx, "level", "debug", "message", "canceling resource")
 
 		return currentState, nil
+	} else if err != nil {
+		return currentState, microerror.Mask(err)
 	}
 
 	var allWorkerInstances []compute.VirtualMachineScaleSetVM